Split APIs by kind
`src/api/index.js` has become too big and APIs are in no particular
order, making it very difficult to work with.

Split the APIs into files per kind so they are easier to work with.
