+++ src/api/utils.js
@@ -1,0 +1,100 @@
+/*
+Copyright 2019-2020 The Tekton Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import { ALL_NAMESPACES } from '@tektoncd/dashboard-utils';
+
+import { getAPIRoot } from './comms';
+
+export const apiRoot = getAPIRoot();
+export const tektonAPIGroup = 'tekton.dev';
+export const triggersAPIGroup = 'triggers.tekton.dev';
+export const dashboardAPIGroup = 'dashboard.tekton.dev';
+
+export function checkData(data) {
+  if (data.items) {
+    return data.items;
+  }
+
+  const error = new Error('Unable to retrieve data');
+  error.data = data;
+  throw error;
+}
+
+export function getAPI(type, { name = '', namespace } = {}, queryParams) {
+  return [
+    apiRoot,
+    '/v1/namespaces/',
+    encodeURIComponent(namespace),
+    '/',
+    type,
+    '/',
+    encodeURIComponent(name),
+    queryParams ? `?${new URLSearchParams(queryParams).toString()}` : ''
+  ].join('');
+}
+
+export function getKubeAPI(
+  type,
+  { name = '', namespace, subResource } = {},
+  queryParams
+) {
+  return [
+    apiRoot,
+    '/proxy/api/v1/',
+    namespace && namespace !== ALL_NAMESPACES
+      ? `namespaces/${encodeURIComponent(namespace)}/`
+      : '',
+    type,
+    '/',
+    encodeURIComponent(name),
+    subResource ? `/${subResource}` : '',
+    queryParams && Object.keys(queryParams).length > 0
+      ? `?${new URLSearchParams(queryParams).toString()}`
+      : ''
+  ].join('');
+}
+
+export function getResourcesAPI(
+  { group, version, type, name = '', namespace },
+  queryParams
+) {
+  return [
+    apiRoot,
+    `/proxy/apis/${group}/${version}/`,
+    namespace && namespace !== ALL_NAMESPACES
+      ? `namespaces/${encodeURIComponent(namespace)}/`
+      : '',
+    type,
+    '/',
+    encodeURIComponent(name),
+    queryParams ? `?${new URLSearchParams(queryParams).toString()}` : ''
+  ].join('');
+}
+
+export function getQueryParams(filters) {
+  if (filters.length) {
+    return { labelSelector: filters };
+  }
+  return '';
+}
+
+export function getTektonAPI(
+  type,
+  { group = tektonAPIGroup, name = '', namespace, version = 'v1beta1' } = {},
+  queryParams
+) {
+  return getResourcesAPI(
+    { group, version, type, name, namespace },
+    queryParams
+  );
+}
