+++ src/api/utils.test.js
@@ -1,0 +1,127 @@
+/*
+Copyright 2019-2020 The Tekton Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import * as utils from './utils';
+
+describe('getAPI', () => {
+  it('returns a URI containing the given type', () => {
+    const uri = utils.getAPI('pipelines');
+    expect(uri).toContain('pipelines');
+  });
+
+  it('returns a URI containing the given type and name', () => {
+    const uri = utils.getAPI('pipelines', { name: 'somename' });
+    expect(uri).toContain('pipelines');
+    expect(uri).toContain('somename');
+  });
+
+  it('returns a URI containing the given type, name, and namespace', () => {
+    const uri = utils.getAPI('pipelines', {
+      name: 'somename',
+      namespace: 'customnamespace'
+    });
+    expect(uri).toContain('pipelines');
+    expect(uri).toContain('somename');
+    expect(uri).toContain('customnamespace');
+  });
+});
+
+describe('getTektonAPI', () => {
+  it('returns a URI containing the given type', () => {
+    const uri = utils.getTektonAPI('pipelines');
+    expect(uri).toContain('pipelines');
+  });
+
+  it('returns a URI containing the given type and name', () => {
+    const uri = utils.getTektonAPI('pipelines', { name: 'somename' });
+    expect(uri).toContain('pipelines');
+    expect(uri).toContain('somename');
+  });
+
+  it('returns a URI containing the given type, name, and namespace', () => {
+    const uri = utils.getTektonAPI('pipelines', {
+      name: 'somename',
+      namespace: 'customnamespace'
+    });
+    expect(uri).toContain('pipelines');
+    expect(uri).toContain('somename');
+    expect(uri).toContain('namespaces');
+    expect(uri).toContain('customnamespace');
+  });
+
+  it('returns a URI without namespace when omitted', () => {
+    const uri = utils.getTektonAPI('clustertasks', {
+      name: 'somename'
+    });
+    expect(uri).toContain('clustertasks');
+    expect(uri).toContain('somename');
+    expect(uri).not.toContain('namespaces');
+  });
+});
+
+describe('getResourceAPI', () => {
+  it('returns a URI containing the given type', () => {
+    const uri = utils.getResourcesAPI({
+      group: 'test.dev',
+      version: 'testversion',
+      type: 'testtype'
+    });
+    expect(uri).toContain('test.dev');
+    expect(uri).toContain('testversion');
+    expect(uri).toContain('testtype');
+  });
+
+  it('returns a URI containing the given type and name without namespace', () => {
+    const uri = utils.getResourcesAPI({
+      group: 'test.dev',
+      version: 'testversion',
+      type: 'testtype',
+      name: 'testname'
+    });
+    expect(uri).toContain('test.dev');
+    expect(uri).toContain('testversion');
+    expect(uri).toContain('testtype');
+    expect(uri).toContain('testname');
+    expect(uri).not.toContain('namespaces');
+  });
+
+  it('returns a URI containing the given type, name and namespace', () => {
+    const uri = utils.getResourcesAPI({
+      group: 'test.dev',
+      version: 'testversion',
+      type: 'testtype',
+      namespace: 'testnamespace',
+      name: 'testname'
+    });
+    expect(uri).toContain('test.dev');
+    expect(uri).toContain('testversion');
+    expect(uri).toContain('testtype');
+    expect(uri).toContain('testname');
+    expect(uri).toContain('namespaces');
+    expect(uri).toContain('testnamespace');
+  });
+});
+
+describe('checkData', () => {
+  it('returns items if present', () => {
+    const items = 'foo';
+    const data = {
+      items
+    };
+    expect(utils.checkData(data)).toEqual(items);
+  });
+
+  it('throws an error if items is not present', () => {
+    expect(() => utils.checkData({})).toThrow();
+  });
+});
