+++ src/containers/Conditions/Conditions.test.js
@@ -19,7 +19,7 @@ import configureStore from 'redux-mock-store';
 import { Route } from 'react-router-dom';
 import { paths, urls } from '@tektoncd/dashboard-utils';
 import { renderWithRouter } from '../../utils/test';
-import * as API from '../../api';
+import * as API from '../../api/conditions';
 import ConditionsContainer from './Conditions';
 
 const namespacesTestStore = {
