+++ src/containers/ServiceAccounts/ServiceAccounts.test.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 The Tekton Authors
+Copyright 2019-2020 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -20,7 +20,7 @@ import thunk from 'redux-thunk';
 import { urls } from '@tektoncd/dashboard-utils';
 import ServiceAccounts from '.';
 import { renderWithRouter } from '../../utils/test';
-import * as API from '../../api';
+import * as API from '../../api/serviceAccounts';
 
 const middleware = [thunk];
 const mockStore = configureStore(middleware);
